#10 Added a new screen to view a video after clicking on it.
- I refactored some of the code as part of that effort to avoid duplicating the
single capsule Redux state. Now we have currentCapsule which is used both for
new capsule creation and viewing the details of a capsule.
- I had heaps of bugs here and there due to spaghetti code in the actions when
saving the capsule. The saved data differed from the Redux state so it'd be
great to re-write the action / db code in a way that prevents such mistakes.