#30 Added the read / unread status on the capsules
- Realm DB config code has been moved to src/config where it has the list of
migrations to perform.
- I added a read attribute to the capsule object with a default value of false
- Capsule is update to read=true when users opens it.

I fear that the current implementation is not clear has I'm updating the read
status in an action called getCapsule - the name of the action doesn't imply
that status will change as well. However I also want to avoid having to overuse
dispatch() in the container. I created issue #31 for it in the backlog.